tests/wpt/tests/css/css-page/cssom/margin-002.html

<!DOCTYPE html>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://drafts.csswg.org/cssom/#the-csspagerule-interface">
<title>Add / remove declarations inside CSSMarginRule</title>
<style id="sheet">
  @page {
    @top-center {}
  }
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  var sheet = document.getElementById("sheet").sheet;
  var marginRule = sheet.rules[0].cssRules[0];
  var style = marginRule.style;

  test(() => {
    assert_equals(style.length, 0);
    style.setProperty("widows", "7");
    assert_equals(style.length, 1);
    assert_equals(style.widows, "7");
    style.setProperty("margin-left", "50%");
    assert_equals(style.length, 2);
    assert_equals(style.marginLeft, "50%");
    style.setProperty("margin-left", "100px");
    assert_equals(style.length, 2);
    assert_equals(style.marginLeft, "100px");
    style.setProperty("margin", "auto");
    assert_equals(style.length, 5);
    assert_equals(style.marginLeft, "auto");
  }, "Add declarations");

  test(() => {
    assert_equals(style.length, 5);
    style.removeProperty("widows");
    assert_equals(style.length, 4);
    style.removeProperty("margin");
    assert_equals(style.length, 0);
  }, "Remove declarations");
</script>
